+++ test/cases/export-default-expression/output.js
@@ -1,4 +1,4 @@
 var exported = (function () {
-const _iife_default = () => {};
+var _iife_default = () => {};
 return _iife_default;
 })();
