+++ test/cases/export-default/output.js
@@ -1,5 +1,5 @@
 var exported = (function () {
-const foo = "123";
+var foo = "123";
 
 return foo;
 })();
