+++ test/cases/export-named-declare-var/output.js
@@ -1,5 +1,5 @@
 var exported = (function () {
-const foo = "123", bar = "456";
+var foo = "123", bar = "456";
 return {
   foo: foo,
   bar: bar
