+++ test/cases/export-named/input.js
@@ -1,2 +1,2 @@
-const foo = "123";
+var foo = "123";
 export {foo as bar};
