+++ test/cases/export-named/output.js
@@ -1,5 +1,5 @@
 var exported = (function () {
-const foo = "123";
+var foo = "123";
 
 return {
   bar: foo
