+++ test/cases/import-conflict/input.js
@@ -1,3 +1,3 @@
 import bar from "foo";
-const foo = "123";
+var foo = "123";
 console.log(bar, foo);
