+++ test/cases/import-conflict/output.js
@@ -1,5 +1,5 @@
 (function () {
 
-const _local_foo = "123";
+var _local_foo = "123";
 console.log(foo, _local_foo);
 })();
