--- test/es5/export-default-expression/input.js
@@ -1,1 +1,0 @@
-export default () => {};
