--- test/es5/export-default-expression/output.js
@@ -1,4 +1,0 @@
-var exported = (function () {
-var _iife_default = () => {};
-return _iife_default;
-})();
