--- test/es5/export-default/input.js
@@ -1,2 +1,0 @@
-var foo = "123";
-export default foo;
