--- test/es5/export-default/output.js
@@ -1,5 +1,0 @@
-var exported = (function () {
-var foo = "123";
-
-return foo;
-})();
