--- test/es5/export-named-declare-var/input.js
@@ -1,1 +1,0 @@
-export var foo = "123", bar = "456";
