--- test/es5/export-named-declare-var/output.js
@@ -1,7 +1,0 @@
-var exported = (function () {
-var foo = "123", bar = "456";
-return {
-  foo: foo,
-  bar: bar
-};
-})();
