--- test/es5/export-named/input.js
@@ -1,2 +1,0 @@
-var foo = "123";
-export {foo as bar};
