--- test/es5/export-named/output.js
@@ -1,7 +1,0 @@
-var exported = (function () {
-var foo = "123";
-
-return {
-  bar: foo
-};
-})();
