--- test/es5/import-conflict-object/input.js
@@ -1,3 +1,0 @@
-import bar from "foo";
-var {foo} = baz;
-console.log(bar, {foo});
