--- test/es5/import-conflict-object/output.js
@@ -1,5 +1,0 @@
-(function () {
-
-var {foo: _local_foo} = baz;
-console.log(foo, {foo: _local_foo});
-})();
