--- test/es5/import-conflict/input.js
@@ -1,3 +1,0 @@
-import bar from "foo";
-var foo = "123";
-console.log(bar, foo);
