--- test/es5/import-conflict/output.js
@@ -1,5 +1,0 @@
-(function () {
-
-var _local_foo = "123";
-console.log(foo, _local_foo);
-})();
